Challenge 20 Submission 4.0
Pet photos condensed to "petFriends" ; submitting because I don't want to touch the "Research Insights" cards for while, the formatting is not perfect but it's the best I've been able to make it because I put another row within the middle 8 columns to have greater control over card placement. I abandoned the code I was writing which would skip the first column of each slide. That was meant to override the margin problems I was having, but I am struggling to apply that idea to the whole page and all other left/right padding I wrote. Instead I tried to clean out all conflicting margins/padding and the alignment of the whole page is not too bad right now. Still a slight horizontal overflow but better than before